requirements.txt
# Core A2A and ADK frameworks
google-adk>=0.1.0
a2a-client>=0.1.0
a2a-server>=0.1.0

# Web framework
starlette>=0.35.0
uvicorn[standard]>=0.25.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0.0

# Validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
